composition-prototype/upscale/requirements.txt
fastapi==0.115.5
transformers==4.46.2
torch==2.3.0
uvicorn==0.32.0
pillow==10.3.0
requests==2.31.0
diffusers==0.31.0
scipy==1.13.0
safetensors==0.4.5